feat: add script and clean up
