+++ .env.template
@@ -1,4 +1,5 @@
 ROOT_DIR = "<enter root directory here>" # run 'pwd' in your terminal and copy contents here
+LOG_LEVEL = "INFO"
 
 OPENFIGI_API_KEY = "<enter open figi api key here>"
 OPENFIGI_URL = "https://api.openfigi.com/v3/mapping"
