+++ .pre-commit-config.yaml
@@ -16,4 +16,5 @@ repos:
         name: isort
         entry: isort
         language: system
+        args: ["--profile", "black"]
         types: [python]
