+++ data/AAPL.json
@@ -1,0 +1,6 @@
+{
+    "isin": "US0378331005",
+    "ticker": "AAPL",
+    "name": "APPLE INC",
+    "esg_report_url": "https://www.apple.com/environment/pdf/Apple_Environmental_Progress_Report_2024.pdf"
+}
